removed comment
This comment was originally left for the language specification team.

At the time of writing this the spec states that entries of
an array/a dictionary are seperated by a postfix comma or a prefix
comma or both if the entries are on seperate lines.
That would have been a nightmare to express in parser combinators
and the parser team chose not to do so.